requirements.txt
astapi
uvicorn[standard]
requests
redis
psycopg2-binary
python-dotenv